LayoutTests/fast/forms/time-multiple-fields/time-multiple-fields-keyboard-events-expected.txt

Multiple fields UI of time input type with keyboard events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Please run this with DumpRenderTree.

Test following keys:
Digit keys
Left/Right - Move focus field inside element
Up/Down - Increment/decrement value of focus field
Tab - Move focus field
Backspace - Make value empty
  
== Digit keys ==
PASS eventsCounter.input is undefined.
PASS eventsCounter.change is undefined.
PASS input.value is "07:56"
PASS eventsCounter.input is 1
PASS eventsCounter.change is 1
== Digit keys starting with zero ==
PASS input.value is "14:03"
== Digit keys and backspace key ==
PASS input.value is "17:06"
== Left/Right keys ==
PASS input.value is "06:05"
PASS document.activeElement.id is "input"
== Up/Down keys ==
PASS input.value is "05:56"
PASS input.value is "03:56"
== Up/Down keys on empty value ==
PASS eventsCounter.input is undefined.
PASS eventsCounter.change is undefined.
PASS input.value is "14:58"
PASS eventsCounter.input is 1
PASS eventsCounter.change is 1
== Tab key ==
PASS input.value is "03:05"
PASS input.value is "07:05"
PASS document.activeElement.id is "another"
== Tab navigation should skip disabled inputs ==
PASS document.activeElement.id is "another"
== Tab navigation should not skip readonly inputs, but editing operations should be ignored. ==
PASS document.activeElement.id is "input"
PASS pseudoOfFocusedSubField(input) is "-webkit-datetime-edit-hour-field"
PASS keyDown("upArrow"); input.value is "01:01"
PASS keyDown("downArrow"); input.value is "01:01"
PASS pseudoOfFocusedSubField(input) is "-webkit-datetime-edit-minute-field"
PASS keyDown("3"); input.value is "01:01"
PASS pseudoOfFocusedSubField(input) is "-webkit-datetime-edit-ampm-field"
PASS pseudoOfFocusedSubField(input) is "-webkit-datetime-edit-minute-field"
PASS document.activeElement.id is "another"
== Shfit+Tab key ==
PASS input.value is "15:00"
PASS input.value is "15:03"
PASS document.activeElement.id is "before"
== Up key on maximum value ==
PASS input.value is "13:00"
== Down key on minimum value ==
PASS input.value is "00:59:59.999"
== Backspace key ==
PASS input.value is ""
== Delete key ==
PASS input.value is ""
== Typeahead ==
PASS input.value is "12:01:56"
PASS input.value is "12:02:56"
== RTL focus navigation ==
The tests in this block fail on platforms without the lang-attribute-aware-form-control-UI feature.
PASS input.value is "01:56"
PASS input.value is "01:02"
PASS input.value is "03:02"
PASS successfullyParsed is true

TEST COMPLETE
